Remove some TODOs
This includes a change to the way the FakeExtractor is set up in the blob
extractor tests:

The previous pattern was to write a test which exercised the production code,
which would not have failed until the production code was written. Production
code being written, the test would then fail due to the inner workings of the
fake, which would complain if you did not set up the behavior in advance.

Passing the test meant modifying the test setup rather than adding production
code.

This change makes it such that you can write a failing test, making an
assertion about what should happen before any production code is written. In
the case of wanting to set up an error case, that functionality is preserved
via the Set..Behavior method, which receives an error in addition to the
normal arguments provided to the fake (as previously). If the error was set,
it will be returned.

A nice side-effect of this is that a bunch of happy-path SetDecompressBehavior
calls are removed, and sad-path SetDecompressBehavior calls are added. This
makes obvious the fact that some tests were failing improperly, because they
didn't satisfy the fake's requirement of having behavior set, rather than
having behavior set that resulted in an error.