Refactors deploy_cmd_test.go to not use a fake DeploymentRecord
In an attempt to increase confidence in the test suite, I am trying to minimize
use of fakes except for the edges of the application.

The deploymentRecord injected into the deployment command under test is now a
real object.  Also all it's dependencies are real except for the FakeFS and a
fake UUID generator as those are the edges.  There are other objects injected
into the deployment command test that are still using fakes but, this pattern
of using real objects except at the edges of the ap seem to be working.
[Finishes #89666128]