Attempts to install golint and go vet into the default GOROOT
rather than installing it into the GoDeps modified GOROOT.  This will cause
those libraries to be installed directly on the machine not in the project
space.  The thought is that those tools should always float at the latest
versions and not be locked down library dependencies of any given project.
This will have the side effect of not dirtying up the project when tests are
run.