Refactor CpiInstaller to take a function instead of returning the installation
We want to be able to clean up installation artifacts when we're done
using the installed CPI -- this will let us keep that cleanup code near
the installation code and not have to worry as much about the cleanup
getting forgotten/refactored out.

This class now has install as a private method and a WithInstall to
run code that needs the cpi installed.

[#95176856](https://www.pivotaltracker.com/story/show/95176856)

Signed-off-by: Tim Hausler <thausler@pivotal.io>