Accept and ignore empty responses.
This patch makes bosh-cli behave like the ruby cli, in that empty
responses are accepted during login but not passed to the identity
provider. This is currently necessary to log in using a local account
when saml is enabled, since uaa prmopts for a login, password, and otp,
but only wants values for the login and password.

[Resolves #59]