Create release generates multi digest sha1/sha256
strings when passed the --sha2 flag.

Updates bosh-utils to include crypto libraries

A non-derterministic configuration hash was removed.
It did not take into account filesystem file ordering.
The configuration hash is not used in the context of create-env,
the vm is always recreated. Therefore the configuration hash is now hard
coded.

The sha1calculator usage was reduced and renamed to digestCalculator.
Usages where sha1s were being verified now only use the utils/crypto
libraries. Remaining usages of the digestCalculator are for digest
string generation. The digestCalculator can be configured to generate
digests for all algorithms provided to the calculator.

[#134621353](https://www.pivotaltracker.com/story/show/134621353)