Sha2 checksumming is a global flag
When the `--sha2` flag is passed, all digests will be produced as
"sha256:asdfasdfasdf" style mulitiple digest checksums in all operations.

Because production of multiple digest strings is a breaking change, it was decided
that sha1 portions of the digest string would be dropped. If a component
is able to parse a multiple digest string, it will choose the strongest
digest anyways, meaning the sha1 will always be ignored.

[#134621327](https://www.pivotaltracker.com/story/show/134621327)