Introduce GCSBlobstore with provider hook
GCSBlobstore introduces initial support for native
GCS blobstore support. We use bosh-gcscli/client as
our client package.
https://github.com/cloudfoundry/bosh-gcscli

This mirrors the existing patterns of S3Blobstore.
Specifically, we leave Delete unimplemented to keep parity.

Significant portions of S3Blobstore are duplicated due to an
SDK-level compatibility issue. Specifically, s3cli/client.Get
expects an io.WriterAt as the SDK accepts that. The GCS SDK
accepts only io.Writer. Hence, we cannot share the code by
having a common interface for bosh-gcscli/client and s3cli/client.
The other option is using interface{} and a type-switch in
each Blobstore method wherever the client is used. The result
of this was undesirable.