When following symlinks, copy the directory permissions of the target
* If we are following symlinks when creating an archive and we encounter
a file that exists inside of a symlinked directory, copy the file
permissions of the target directory when creating the new directory.

* We also noticed that these tests will run differently depending on the
umask value of the system running the tests. To make sure the file
permissions are always being set consistently regardless of the
environment the tests are run on, we manually set the umask to be 0
before each test.

[#145780425]

Signed-off-by: Eric Promislow <eric.promislow@suse.com>