Test for source when uploading a release
When uploading releases, the bosh-cli first tests if a release exists
before attempting to upload that release.  Previously, it would consider
a release to exist on the director regardless of whether source bits
exist for that release.  This could result in skipping release uploads
when they actually should be uploaded in the case where the only
releases that are present are compiled.  This commit changes that in the
following way: when checking whether a release has already been uploaded
to the director and no stemcell is specified, we consider if the given
release has any packages that have both a sha1 and a blobstore_id and
therefore have source bits that can be compiled when needed.

[#161936793](https://www.pivotaltracker.com/story/show/161936793)