Ignore missing local file if release exists
If the release is already present on the director, and an operator
attempts to deploy with the file missing, then the file will be ignored
and the existing release will be used. If the file is present and can be
read, it will take precendence over an existing release on the director.

[#166850595](https://www.pivotaltracker.com/story/show/166850595)