Check if the release needs to be uploaded only if the upload fails
This change moves the logic for checking if the release needs to be
uploaded back to being after attempting to upload the release. This
change is necessary due to the fact that we are currently unable to
distinguish between source and compiled releases when going through the
upload flow.

With the previous logic, if you were to upload a compiled release and
then a source release of the same version, the source bits would never
be uploaded.

[#166850595](https://www.pivotaltracker.com/story/show/166850595)

Co-authored-by: Rebecca Putinski <rputinski@pivotal.io>