`inspect-local-release` now cleans up temp files
Invoking ArchiveReader.Read() returns a Release object instance that -as
part of its instantiation- has created files in `$HOME/.bosh/tmp`. It's
up to the caller to call `Release.CleanUp()` when they are done with the
object.

This commit fixes one place where that call to CleanUp was not performed.

I know that there are _other_ places where we are failing to clean up,
but I do not yet know where they are.

NOTE: Tests have not been added to ensure that CleanUp continues to be
      called. The tests for InspectLocalReleaseCmd are substantially
      faked out, and I don't want to think about how to update the
      integration tests to test the `i-l-r` command, given that they
      don't do that _now_.

[#168201650] Bosh cli writes to .bosh/tmp and never cleans up after itself so customer disks fill up