Use default formating for progress bar rtime
The remaining time template of pb.ProgressBar requires a '%s' for both
the active format and the completed format. When an empty string is
provided for the completed format, it will attempt to Sprintf that
format with the final elapsed time. This causes the progress bar to
print something like:
> [-----------------------------------------------] 100.00%    %!(EXTRA string=0s)

See the pb.ProgressBar source:
https://github.com/cheggaaa/pb/blob/3058580baeb74f7bfc7e9ca1b924f34cc070f0fd/v3/element.go#L285

By using the default format, the "remaining" time will be printed while
the upload / download is active, and then the final elapsed time will be
printed when completed.

[#178934051] Likely incorrect `fmt.Printf` somewhere in the code path for `bosh deploy`

Signed-off-by: Brian Upton <bupton@vmware.com>