If using wsl2 virtualbox cant be installed on the linux vm itself
It needs to run on the windows host. The create env would fail because:
The cmd Structs UseIsolatedEnv field was hardcoded to true and prevented access to binaries outside the hardcoded PATH. That means that the Windows Native VBoxManage binary could not be accessed by the CPI.

For the feature to fully work it requires an additional PR to the vbox cpi. The CPI shells out to the VBoxManage Binary to create a network and uses a "name" parameter. On Windows that value would contain " "(spaces), thus the call failed with too many parameters error on the VBoxManage Binary.

These PRs fix these issues and allow running the virtualbox cpi on wsl2