Multiline String Var Handling behaviour
the fix in aba747f165f039fbe24b4dba18baa142e6e96f2d introduced changing
for values surrounded by single and double quotes. Previously these would
be dropped for STRING values and kept for bool/int/float (the types yaml knows about)

if the value in question is a string (and not one of the other YAML types) the new IF statement circumvents YAML.unmarshal value and uses the value from the array. This will still contain the quotes if they were used within the command line.

examples of changed behaviour
https://ci.bosh-ecosystem.cf-app.com/teams/main/pipelines/bosh-director/jobs/compile-bosh-dev-release-bionic/builds/73

vs

https://ci.bosh-ecosystem.cf-app.com/teams/main/pipelines/bosh-director/jobs/compile-bosh-dev-release-bionic/builds/72