🐞 Deprecate compiled pkg deps, like Director
The BOSH Director no longer requires the dependencies of compiled
packages to be included the release (the dependencies have served their
purpose: the compilation of the package. Now that the package has been
compiled, the dependencies can be discarded like used tissue paper. This
helps shrink the size of exported releases).

This commit brings the BOSH CLI in conformance with the Director, no
longer requiring the dependencies of compiled packages.

Fixes, on the BOSH Director integration tests:
```
Constructing release from manifest:
  - Constructing compiled packages from manifest:
      - Expected to find compiled package 'b' since it's a dependency of compiled package 'a'
```