Fix bug in uploading compiled releases
If the bosh director had only source packages, and you were trying
to upload a compiled release, it would not upload the release.

This behavior was only seen when the release was being uploaded
as a local file referenced via url: file:// in the deployment
manifest.

The behavior of ReleaseHasCompiledPackage is now back to the original,
it will return true if the director has ANY package for the release
compiled for the given stemcell. We decided not to check for ALL packages
in the release, because it is possible you may have some packages
in a release that never get compiled for a particular OS.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>