When deciding if to upload compiled releases, check for ALL packages on the director
Previously we were assuming that if the director had a single package compiled for
the stemcell we have, we did NOT need to upload the compiled release.

This is not always true. Because packages can be reused across multiple versions of a release.
You can upload a release R1 that has packages A1 and B1 all compiled for stemcell S1,
then you upload release R2 that has packages A1 and B2 and compiled for stemcell S2.

If you then try to upload a compiled release R1 for stemcell S2 we inspect the director
and find that we have A1 already compiled for S2, and skip the upload, but we don't
have B1 compiled for S2 and the deployment fails.

[#181275623]

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>