Stop using `gopath` in CI, harmonize GOARCH suffix
Overall we are removing pre-`go.mod` Golang patterns and dealing with
the resulting fallout.

We no longer need to create a GOPATH style directory layout to build
golang projects so we are removing that.

Simplify build script and artifact handling by making all builds use a
GOARCH suffix, not just `arm64`.

Add `set -eu -o pipefail` to all builds.

Fix homebrew formula creation so that we include a shasum for
darwin-arm64.