Adds --system and --all-logs flags to logs action
This commit adds the --system and --all-logs flags to the "bosh logs"
action.

The --system flag allows an operator to fetch or follow logs in the
'/var/log' directory in a Bosh-deployed VM. The --all-logs flag allows
an operator to fetch or follow logs provided by all of the --jobs (the
default flag if none are specified), --agent, and --system flags in one
convenient operation.

The code for the --system flag in the --follow operation adds a
dependency on the 'find' command, but we believe that we currently
include that command in every stemcell that we ship, so this should not
be a problem. Because /var/log typically contains rotated logs, as well
as files that are binary data, we had file-exclusion requirements that
were too complex to be expressed with a glob pattern.

This commit also contains a rename of the "job" function parameter in
Client.FetchLogs to "instance". The name for this concept has been
"instance" for at _least_ five years, and this discrepancy was _Very
Confusing_ to us when we were trying to understand that section of the
code. So, we renamed the variable.

It's also important to note that use of the --follow flag should work
with any version of the Bosh Agent, but _downloading_ the --system or
--all-logs logs (done by omitting the --follow flag) requires Bosh Agent
changes to handle the new log types. This change is backwards
compatible- incompatible Agents will return an error if either the
--system or --all-logs flag is used, and will work correctly if neither
one is used.

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>