Add ability for retrieving logs from directors
The `bosh logs` command now has a `--director` flag that works in conjunction with the
`--agent-endpoint` and `--agent-certificate` flags to talk directly to the director's agent HTTP endpoint.
All of the standard log features (`--follow`, `--agent`, etc.) should
work the same as they do with the standard `bosh logs` command.

(This ability also applies to other VMs created using the `bosh create-env` command)

[#183962473](https://www.pivotaltracker.com/story/show/183962473)

Authored-by: Chris Selzo <cselzo@vmware.com>