Add ability to scp to/from directors
The `bosh scp` command now has a `--director` flag that works in conjunction with the
`--agent-endpoint` and `--agent-certificate` flags to talk directly to the director's agent HTTP endpoint.
All of the standard scp features (`--recursive` etc.) should
work the same as they do with the standard `bosh scp` command.

Note that using the `--private-key` flag is not currently supported in
conjunction with the `--director` flag

(This ability also applies to other VMs created using the `bosh create-env` command)

[#183962473](https://www.pivotaltracker.com/story/show/183962473)

Co-authored-by: Joseph Palermo <jpalermo@vmware.com>
Co-authored-by: Chris Selzo <cselzo@vmware.com>