Auto-release CLI on new version of golang
Adds task to the `automatically-release-new-patch` job that compares the
golang version of the previously released BOSH CLI binary with the
golang version of what would be released. If they're different (in the
case of a bumped golang version, or intentional roll back), then an
automatic release would be triggered, similar to when CVE fixes are
detected.