Remove PowerDNS dependent `--dns` flang ad related functionality
PowerDNS is being removed from the BOSH Director, and as a consequence
the data returned from the `instances` and `vms` commands no longer have
DNS data returned.

This commit removes the `--dns` flag and logic to display the data
returnd when it is sent.

Add a spec to validate that the CLI continues to handle responses from
directors which still have PowerDNS code. Concretely this means the VM
info response has a `"dns"` key in it.

See: https://github.com/cloudfoundry/bosh/pull/2501