Update auto-release pattern
Previously, if the patch version of the CLI was bumped due to the
`check-for-updated-binary-version` task returning successfully, then the
`check-for-patched-cves` task afterward would fail because it would try
to look up the new version as a git tag. If that didn't cause a failure
and it also returned successfully, the version could have been bumped
twice.

Now, the golang-release task scripts create a
release-notes/needs-release file when the condition for the script is
met. This allows us to chain the scripts together, build up release
notes from each script, then bump a single time if any of the scripts
were successful.