Allow usage of pre-compiled CPI for create-env command
The machinery to allow the use of pre-compiled releases for things
_other_ than the CPI has existed for some years.

Note that the CPI must be compiled for the architecture and OS of the
machine you are running the create-env command from.  At the time of
writing, we're not sure if that architecture and OS must match that of
the BOSH director you're creating.

[#187692170]

Co-authored-by: Chris Selzo <chris.selzo@broadcom.com>