Add an explicit fetch_tags
The check-for-patched-cves task compares the current commit with the tag
of the last released version. Due to concourse caching, you can end up
with a version of the bosh-cli repo resource that is newer than the last
release tag, but does not have the tag.

Adding fetch_tags to the get for the resource _should_ force Concourse
to get any tags added to the repo since the commit first started flowing
through the pipeline.