Unset BOSH_ALL_PROXY for the duration of the integration tests
This value will be captured and re-exported after the test run.

If you attempt to run these tests, which mock out calls to the BOSH
director,  while BOSH_ALL_PROXY is set, the bosh cli will respect that
setting and try to use your targetted director, which is not ideal!