Fix mockgen script
This change fixes mockgen by changing the package name to the reflect
the module name github.com/cloudfoundry/bosh-cli/v7. Presumably, the
module name changed at some point, but the mockgen script wasn't updated
to reflect it. So it would try to add github.com/cloudfoundry/bosh-cli
as a go mod dependency and complain vendoring is inconsistent (you can't
vendor bosh-cli in its own repo).

We will be regenerating mocks using the updated script, so it's good to
have it in working order.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>