Remove pgzip and use tar directly
This is basically a revert of
https://github.com/cloudfoundry/bosh-cli/pull/669.

We found that the bosh cli with the pgzip code produces releases that
cannot be compiled by the bosh agent.  In addition, there were many many
edge cases that we found we had to fix and/or paper over that have been
solved in the tar executable over the past decades.

Maybe in the future we can give this another go, as the performance
improvements would be welcome.

Signed-off-by: Ming Xiao <ming.xiao@broadcom.com>