Ruby TemplateEvaluationContext handles missing ERB
This change updates the exception catching in the embeded ruby script
(template_evaluation_context.rb) so that when the supplied ERB template
argument is not a file on disk the exception handling code does not
itself error.

Prior to this change if the ruby code attempted to load a non-existent
ERB template the exception handling code would fail tyring ot access the
ERB instance's `#filename` method.